requirements-py3.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
configparser==3.5.0
decorator==4.3.0
future==0.17.1
gaphas==1.0.0rc1
isort==4.3.4
jsonconversion==0.2.9
lazy-object-proxy==1.3.1
mccabe==0.6.1
psutil==5.7.0
pycairo
PyGObject
pylint==1.9.3
python-gtkmvc3-dlr==1.0.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
wrapt==1.10.11
yaml-configuration>=0.1.7
